pyproject.toml
[project]
name = "agentic_ai"
version = "1.0.0"
description = "AI-powered agents for content generation, analysis, and automation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    "python-pptx>=1.0.0",
    "xhtml2pdf>=0.2.15",
    "fpdf2>=2.8.0",
    "requests>=2.32.0",
    "replicate>=1.0.0",
]

[project.scripts]
agentic-ai = "app.main:main"

